TEST-#2951: use conda for base env
