+++ asv_bench/benchmarks/scalability/scalability_benchmarks.py
@@ -17,6 +17,7 @@ from modin.pandas.utils import from_pandas
 try:
     from modin.utils import to_pandas
 except ImportError:
+    # This provides compatibility with older versions of the Modin, allowing us to test old commits.
     from modin.pandas.utils import to_pandas
 import pandas
